started implementing upvalues
